backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
lxml==4.9.3
requests==2.31.0